requirements.txt
Django==3.0.3
django_polymorphic==2.1.2
